@@ -4,26 +4,47 @@
   "description": "Serverless aws-nodejs-typescript template",
   "main": "serverless.ts",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "dev": "nodemon app.ts",
+    "build": "rimraf dist && tsc"
   },
   "engines": {
     "node": ">=14.15.0"
   },
   "dependencies": {
     "@middy/core": "^3.4.0",
-    "@middy/http-json-body-parser": "^3.4.0"
+    "@middy/http-json-body-parser": "^3.6.2",
+    "@types/cors": "^2.8.13",
+    "@types/express-jwt": "^7.4.2",
+    "aws-serverless-express": "^3.4.0",
+    "cors": "^2.8.5",
+    "dotenv": "^16.0.3",
+    "express": "^4.18.2",
+    "express-jwt": "^8.4.1",
+    "http": "^0.0.1-security",
+    "jwks-rsa": "^3.0.1",
+    "morgan": "^1.10.0",
+    "nodemon": "^2.0.22",
+    "pg": "^8.10.0",
+    "serverless-http": "^3.2.0"
   },
   "devDependencies": {
     "@serverless/typescript": "^3.0.0",
     "@types/aws-lambda": "^8.10.71",
-    "@types/node": "^14.14.25",
+    "@types/aws-serverless-express": "^3.3.5",
+    "@types/express": "^4.17.17",
+    "@types/morgan": "^1.9.4",
+    "@types/node": "^14.18.42",
+    "@types/pg": "^8.6.6",
     "esbuild": "^0.14.11",
     "json-schema-to-ts": "^1.5.0",
+    "rimraf": "^5.0.0",
     "serverless": "^3.0.0",
     "serverless-esbuild": "^1.23.3",
-    "ts-node": "^10.4.0",
+    "serverless-offline": "^12.0.4",
+    "ts-node": "^10.9.1",
     "tsconfig-paths": "^3.9.0",
-    "typescript": "^4.1.3"
+    "typescript": "^4.9.5"
   },
   "author": "The serverless webpack authors (https://github.com/elastic-coders/serverless-webpack)",
   "license": "MIT"
