@@ -6,9 +6,9 @@ import schema from './schema';
 
 const hello: ValidatedEventAPIGatewayProxyEvent<typeof schema> = async (event) => {
   return formatJSONResponse({
-    message: `Hello ${event.body.name}, welcome to the exciting Serverless world!`,
+    message: `Hello, welcome to the exciting Serverless world!`,
     event,
   });
 };
 
-export const main = middyfy(hello);
+export const main = middyfy(hello)
